total assign

+++ src/main/java/shupship/dto/ReportAllDepts.java
@@ -11,19 +11,19 @@ public class ReportAllDepts {
 
     private String postCode;
 
-    private BigInteger employees;
+    private Long employees;
 
-    private BigInteger totalAssigns;
+    private Long totalAssigns;
 
-    private BigInteger leads;
+    private Long leads;
 
-    private BigInteger successes;
+    private Long successes;
 
-    private BigInteger contacting;
+    private Long contacting;
 
-    private BigInteger fails;
+    private Long fails;
 
-    private BigInteger employeeNotAssigned; // SL cá nhân chưa có tiếp xúc
+    private Long employeeNotAssigned; // SL cá nhân chưa có tiếp xúc
 
-    private BigInteger assigned;
+    private Long assigned;
 }
